+++ src/AuthorityController.java
@@ -1,4 +1,3 @@
-
 /**Handles all interaction between the associated model class and the view class.*/
 public class AuthorityController {
 	
@@ -28,7 +27,13 @@ public class AuthorityController {
 	//get information from model, and set labels, etc in view
 	public void initView()
 	{
-		
+		view.getWelcomeLabel().setText("Hello, "+ model.getName() );
+		view.getDepNumLabel().setText("# of Departments: "+ model.getDepNum() );
+		model.getStats();
+		view.getPatientTotalLabel().setText("# of Registered Patients: "+model.getPatientNum());
+		view.getDoctorTotalLabel().setText("# of Registered Doctors: "+model.getDoctorNum());
+		view.getNurseTotalLabel().setText("# of Registered Nurses: "+model.getNurseNum());
+
 	}
 	
 	//initialize the listeners from the view class that need to interact with model